move screenshot method into mixin
